This EA is a telegram bot that listens for user commands.
The chart scanning is done through a CAbtSrsiRatio
class, which we declare as an array.

The processing and sending of messages is done
through a CCustomBot class, which we inherit
and then override the function ProcessMessages().

In the ProcessMessages() function, we loop through
our array, and call the strPrintTelegram() function
of each instance.

For each symbol and timeframe pair, e.g EURUSD D1,
we create an instance of the class CAbtSrsiRatio.
We call the Init() function and pass it the symbol
and timeframe. And finally, we add the instance to
our array of objects.

In the OnTimer() event of the EA, we loop through
our array, and call the IndicatorValues() function
of each instance.

In the OnDeinit() event of the EA, we loop through
our array, and call the DeInit() function of each
instance.